Code cleanup: comments and refactoring

+++ board.h
@@ -68,7 +68,7 @@ void undoMove(Board* board);
     int validMove(uint64 move);
     void printPieces(const Board* board);
     void printBoard(const Board* board);
-    void getMoveString(int move, char* moveString);
+    void getMoveString(uint64 move, char* moveString);
     void printBitboard(uint64 bitboard);
 #endif
